+++ lib/Elastica/QueryBuilder/DSL/Collapse.php
@@ -2,7 +2,7 @@
 
 namespace Elastica\QueryBuilder\DSL;
 
-use Elastica\Query\InnerHits;
+use Elastica\Collapse\InnerHits;
 use Elastica\QueryBuilder\DSL;
 
 /**
@@ -12,11 +12,17 @@ use Elastica\QueryBuilder\DSL;
  */
 class Collapse implements DSL
 {
+    /**
+     * @return string
+     */
     public function getType(): string
     {
         return self::TYPE_COLLAPSE;
     }
 
+    /**
+     * @return InnerHits
+     */
     public function inner_hits(): InnerHits
     {
         return new InnerHits();